backend/requirements.txt
# Fantasy League Chatbot - Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Fuzzy Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# HTTP Requests
httpx==0.25.1
requests==2.31.0
cloudscraper==1.2.71

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/LLM
openai==1.3.7
anthropic==0.7.0

# Google Sheets
gspread==5.12.0
google-auth==2.23.4

# Environment
python-dotenv==1.0.0

# Excel reading
openpyxl==3.1.2

# Testing (optional for now)
pytest==7.4.3
pytest-asyncio==0.21.1
